size_t fixes.

+++ code/3DSHelper.h
@@ -497,7 +497,7 @@ struct Node
 
 	~Node()
 	{
-		for (unsigned int i = 0; i < mChildren.size();++i)
+		for (size_t i = 0; i < mChildren.size();++i)
 			delete mChildren[i];
 	}
